requirements.txt
Flask==2.3.1
PyYAML==6.0
flask-RESTful==0.3.9
singleton-decorator==1.0.0
requests==2.28.1
